requirements.txt
flask==3.1.0
flask-cors==5.0.1
feedparser==6.0.11
requests==2.32.3
beautifulsoup4==4.13.3
google-generativeai==0.8.5
gunicorn==23.0.0
python-dateutil==2.9.0
